Refactoring to update balance on account service

+++ src/main/java/com/naiomi/paymentservice/account/AccountDto.java
@@ -3,8 +3,8 @@ package com.naiomi.paymentservice.account;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public record AccountDto(String accountId, int balance) {
-    public int getBalance() {
+public record AccountDto(String accountId, double balance) {
+    public double getBalance() {
         return balance;
     }
 }
